% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bartlett_uni.R
\name{bartlett_uni}
\alias{bartlett_uni}
\title{Scaled Battlett kernel}
\usage{
bartlett_uni(v, b)
}
\arguments{
\item{v}{Input value.}

\item{b}{Bandwidth.}
}
\description{
This function calculates the scaled kernel \eqn{K_b(v) = b^{-1}K(v/b)} where
\eqn{K(v) = \max\{1 - |v|\}.}
}
