% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohByFreq.R
\name{CohByFreq}
\alias{CohByFreq}
\title{(Partial) coherence for a given frequency}
\usage{
CohByFreq(
  w1,
  w2,
  sp.est,
  type = "partial",
  i = NULL,
  j = NULL,
  sp.IR = NULL,
  H.list = NULL,
  ppp = NULL
)
}
\arguments{
\item{w1, w2}{Frequency vector (only allow frequency values evaluated in \code{sp.est})}

\item{sp.est}{A list of kernel spectral estimate matrices of the point pattern.}

\item{type}{If \code{type = 'partial'}, compute partial coherence. If \code{type = 'normal'},
compute coherence.}

\item{i, j}{Optional. Index of the multivariate point process.}

\item{sp.IR}{Optional. Spectrum estimate of the intensity reweighted process, which is
calculated by \code{\link[=IRspec]{IRspec()}}. If this argument is specified, then \code{w1}, \code{w2}, \code{H.list},
\code{ppp} are not required.}

\item{H.list}{Optional. A list from \code{\link[=Hmatrix]{Hmatrix()}}. This argument is only required
if \code{sp.IR} is unspecified.}

\item{ppp}{Optional. Point pattern. This argument is only required if \code{H.list}
is unspecified.}
}
\value{
If both \code{i} and \code{j} are not specified (default), return a \eqn{m \times m}
matrix. If they are specified, return a value.
}
\description{
For a frequency \eqn{\boldsymbol\omega}, compute its squared magnitude coherence
\eqn{\hat{R}^{(a,b)}(\boldsymbol\omega)} and partial coherence \eqn{\hat{D}^
{(a,b)}(\boldsymbol\omega)} between two point processes \eqn{X_a} and \eqn{X_b}
as follows:
\deqn{\hat{R}^{(a,b)}(\boldsymbol\omega) = \frac{|\hat{F}^{(a,b)}(\boldsymbol
\omega)|^2}{\hat{F}^{(a,a)}(\boldsymbol\omega)\hat{F}^{(b,b)}(\boldsymbol\omega)}
\quad \text{and} \quad
\hat{D}^{(a,b)}(\boldsymbol\omega) = \frac{|\hat{F}^{-(a,b)}(\boldsymbol
\omega)|^2}{\hat{F}^{-(a,a)}(\boldsymbol\omega)\hat{F}^{-(b,b)}(\boldsymbol\omega)},}
where \eqn{\hat{F}^{-(a,b)}} is the \eqn{(a,b)}th element of the inverse spectrum
estimator \eqn{\hat{F}(\boldsymbol\omega)^{-1}}.
}
