% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixpd_fun.R
\name{fixpd_fun}
\alias{fixpd_fun}
\title{Fix the positive definite issue for a matrix}
\usage{
fixpd_fun(mat)
}
\arguments{
\item{mat}{A complex-valued matrix.}
}
\description{
Fix the positive definite issue for a matrix
}
