% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cate.R
\name{generate_cate}
\alias{generate_cate}
\title{Generate all pairwise combinations of mark}
\usage{
generate_cate(cate)
}
\arguments{
\item{cate}{Vector containing the categorical mark of a point process, which
can be obtained from \code{levels(spatstat.geom::marks(ppp))}, where \code{ppp} is the
point pattern.}
}
\value{
A matrix.
}
\description{
Generate all pairwise combinations of mark
}
