% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H.h.lambda.1.R
\name{H.h.lambda.1}
\alias{H.h.lambda.1}
\title{Compute H.h.lambda.1 in the centered DFT}
\usage{
H.h.lambda.1(w2, w1, a, taper, A1, A2, inten.fitted)
}
\arguments{
\item{w1, w2}{A numeric value or vector of frequency values at horizontal and
vertical directions, respective.}

\item{a}{Taper coefficient, a value within unit interval. If \code{a = 1}, then
taper is not applied.}

\item{taper}{Data taper function.}

\item{A1, A2}{Side lengths of the observation window.}

\item{inten.fitted}{Intensity function of individual point pattern.}
}
\description{
Compute H.h.lambda.1 in the centered DFT
}
