% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRspec.R
\name{IRspec}
\alias{IRspec}
\title{Compute the spectrum of the intensity reweighted process}
\usage{
IRspec(w1, w2, sp.est, i = NULL, j = NULL, H.list = NULL, ppp)
}
\arguments{
\item{w1, w2}{Frequency vector (only allow frequency values evaluated in \code{sp.est})}

\item{sp.est}{A list of kernel spectral estimate matrices of the point pattern.}

\item{i, j}{Optional. Index of the multivariate point process.}

\item{H.list}{Optional. The list from \code{Hmatrix()}.}

\item{ppp}{Point pattern.}
}
\description{
Compute the spectrum of the intensity reweighted process
}
