% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_cate.R
\name{pair_cate}
\alias{pair_cate}
\title{Generate all pairwise combinations of}
\usage{
pair_cate(cate)
}
\arguments{
\item{cate}{Vector containing the categorical mark of a point process, which
can be obtained from \code{levels(spatstat.geom::marks(ppp))}, where \code{ppp} is the
point pattern.}
}
\description{
Generate all pairwise combinations of
}
