% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{periodogram}
\alias{periodogram}
\title{Periodogram of a univatiate point pattern}
\usage{
periodogram(
  i,
  j,
  ppp,
  inten.formula = "~1",
  data.covariate = NULL,
  a = 0.025,
  return.DFT = FALSE,
  A1 = NULL,
  A2 = A1,
  ext.factor = NULL,
  endpt = 1.5,
  generate_freq. = generate_freq,
  taper. = taper,
  H.h.lambda.1. = H.h.lambda.1
)
}
\arguments{
\item{i}{Mark index. An element in \code{levels(spatstat.geom::marks(ppp))}.}

\item{j}{Mark index. An element in \code{levels(spatstat.geom::marks(ppp))}.}

\item{ppp}{A point pattern of class \code{"ppp"}.}

\item{inten.formula}{A \code{\link{formula}} syntax in character format specifying the
log-liner model for the intensity function, which is passed to \code{\link[=ppm]{ppm()}}. The
default is constant intensity \code{inten.formula = "~1"}.}

\item{data.covariate}{Optional. The values of spatial covariates passed to
the \code{data} argument in \code{\link[=ppm]{ppm()}}.}

\item{a}{Taper coefficient, a value within \eqn{(0,1/2)}. If \code{a = 0}, then
taper is not applied. Default is \code{a = 0.025}.}

\item{return.DFT}{If \code{TRUE}, also return the centered discrete Fourier
transform (DFT).}

\item{A1, A2}{Optional. Side lengths of the observation window.}

\item{ext.factor}{Optional. If \code{NULL} (default), the frequency grid for the
periodogram is not extended. Please keep this \code{NULL} unless you know what you
are doing.}

\item{endpt}{A positive value indicating the scale factor of the endpoint frequency.}

\item{generate_freq., taper., H.h.lambda.1.}{Functions in their respective .R files.}
}
\value{
A matrix if \code{return.DFT = FALSE}; a list consisting of the periodogram and the
centered DFT if \code{return.DFT = TRUE}.
}
\description{
Given a multivariate point pattern \code{ppp}, compute the periodogram for each
marginal point pattern.
}
