% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{periodogram_smooth}
\alias{periodogram_smooth}
\title{Kernel spectral estimator of a multivatiate (multitype) point pattern}
\usage{
periodogram_smooth(
  ppp,
  i = NULL,
  j = i,
  inten.formula = "~1",
  data.covariate = NULL,
  bandwidth,
  correct = TRUE,
  a = 0.025,
  A1 = NULL,
  A2 = A1,
  endpt = 1.5,
  equal = FALSE,
  kern = bartlett_uni
)
}
\arguments{
\item{ppp}{A point pattern of class \code{"ppp"}.}

\item{i}{Mark index. An element in \code{marks(ppp)}.}

\item{j}{Mark index. An element in \code{marks(ppp)}.}

\item{inten.formula}{A \code{\link{formula}} syntax in character format specifying the log-liner model for the
intensity function, which is passed to \code{\link{ppm}}. The default is constant intensity \code{inten.formula = "~1"}.}

\item{data.covariate}{Optional. The values of spatial covariates passed to the \code{data} argument in \code{\link{ppm}}.}

\item{bandwidth}{A positive value indicating the bandwidth of kernel, determined by \code{\link[=select_band]{select_band()}}.}

\item{correct}{Logical. If \code{TRUE} (default), conduct edge correction when computing the kernel spectral estimator.}

\item{a}{Taper coefficient, a value within unit interval. If \code{a = 1}, then no data taper is used.
Default is \code{a = 0.025}.}

\item{A1, A2}{Optional. Side lengths of the observation window.}

\item{endpt}{A positive value indicating the scale factor of the endpoint frequency.}

\item{equal}{Logical. If \code{TRUE}, then use the same bandwidth for both x and y direction. Default is \code{FALSE}.}

\item{kern}{Univariate scaled kernel function, e.g., Barrlett kernel (default).}
}
\value{
A list of matrices, or a single matrix if \code{i} and \code{j} are specified.
}
\description{
Computes the kernel spectral estimator for a multivatiate (multitype) point pattern.
}
\details{
The minimal required arguments are \code{ppp}, \code{inten.formula}, and \code{bandwidth}.
If you use any spatial covariate (other than the Cartesian coordinates) in \code{inten.formula}, then
\code{data.covariate} is also needed. All the other arguments can be left by default setting.
\code{periodogram_smooth()} computes all the pairwise (marginal and cross-) kernel spectral estimators
automatically when the mark indices \code{i} and \code{j} are unspecified. If \code{i} and \code{j} are specified,
then it only computes the result for that mark combination.

The bandwidth can be determined by \code{\link[=select_band]{select_band()}}.
}
\examples{
library(spatstat)
lam <- function(x, y, m) {(x^2 + y) * ifelse(m == "A", 2, 1)}
spp <- rmpoispp(lambda = lam, win = square(5), types = c("A","B"))

# Compute kernel spectral estimator with intensity fitted by log-linear model
# with Cartesian coordinates
periodogram_smooth(spp, inten.formula = "~ x + y", bandwidth = 1.2)
}
