% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pairs.R
\name{plot_pairs}
\alias{plot_pairs}
\title{Plot all possible spectral density estimates}
\usage{
plot_pairs(est.list, ppp, xnorm = TRUE, type = "Re", shared.legend = TRUE)
}
\arguments{
\item{est.list}{List. The kernel spectral density estimate from \code{\link[=periodogram_smooth]{periodogram_smooth()}}.}

\item{ppp}{A point pattern of class \code{"ppp"}.}

\item{xnorm}{Logical. If \code{TRUE} (default), plot the radially-averaged spectral
estimates. Otherwise, plot the raw values by heatmap.}

\item{type}{If \code{type = "Re"} (default), plot the real part of the estimates.
If \code{type = "Im"}, plot the imaginary part.}

\item{shared.legend}{Logical. Whether to share the legend across all plots.}
}
\description{
Given a multivariate (multitype) point pattern, use \code{ggplot2} package to
visualize all (marginal and cross) spectral density estimates.
}
\examples{
library(spatstat)
lam <- function(x, y, m) {(x^2 + y) * ifelse(m == "A", 2, 1)}
set.seed(227823)
spp <- rmpoispp(lambda = lam, win = square(5), types = c("A","B"))
KSDE.list <- periodogram_smooth(spp, inten.formula = "~ x + y", bandwidth = 1.15)

plot_pairs(est.list = KSDE.list, ppp = spp)
plot_pairs(est.list = KSDE.list, ppp = spp, type = "Im")
plot_pairs(est.list = KSDE.list, ppp = spp, xnorm = FALSE)
plot_pairs(est.list = KSDE.list, ppp = spp, xnorm = FALSE, type = "Im")
}
