% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual.R
\name{plot_period}
\alias{plot_period}
\title{Plot individual spectral estimate}
\usage{
plot_period(
  period,
  freq.list = NULL,
  remove = NULL,
  title = NULL,
  palette = "Spectral",
  legend.range = NULL
)
}
\arguments{
\item{period}{Matrix. Output from \code{periodogram()} or \code{\link[=periodogram_smooth]{periodogram_smooth()}}.}

\item{freq.list}{List. Output from \code{\link[=generate_freq]{generate_freq()}} with \code{return.comb = TRUE}.}

\item{remove}{An experimental feature. Please ignore.}

\item{title}{Character. The title of the plot.}

\item{palette}{Color schemes for visualization. Type \code{?scale_fill_distiller()}
to check available options.}

\item{legend.range}{Input \code{c(lower, upper)} to manually set the lower and upper limit of the spectral density to visualize.}
}
\description{
Use \code{ggplot2} package to plot the heatmap of a single periodogram or kernel spectral estimate.
}
\examples{
library(spatstat)
lam <- function(x, y, m) {(x^2 + y) * ifelse(m == "A", 2, 1)}
spp <- rmpoispp(lambda = lam, win = square(5), types = c("A","B"))
period.list <- periodogram_smooth(spp, inten.formula = "~ x + y", bandwidth = 1.2)
names(period.list)

# Only plot the real part of the periodogram
period.list <- lapply(period.list, Re)
freq.list <- generate_freq(A1 = diff(spp$window$xrange), A2 = diff(spp$window$yrange),
                           return.comb = TRUE)
plot_period(period.list[[1]], freq.list = freq.list,
            title = "Kernel spectral estimate: type A")
plot_period(period.list[[2]], freq.list = freq.list,
            title = "Kernel spectral estimate: cross-spectrum")
plot_period(period.list[[3]], freq.list = freq.list,
            title = "Kernel spectral estimate: type B")
}
\seealso{
\code{\link[=plot_pairs]{plot_pairs()}}
}
