% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale2unitarea.R
\name{scale2unitarea}
\alias{scale2unitarea}
\title{Recale the rectangular image to unit square}
\usage{
scale2unitarea(im, length.x, length.y, return.fun = TRUE)
}
\arguments{
\item{im}{Image object.}

\item{length.x, length.y}{Side lengths.}

\item{return.fun}{Logical. If \code{TRUE}, convert the scaled image to function.
Otherwise, return the image itself.}
}
\description{
Recale the rectangular image to unit square
}
