% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwidth.R
\name{select_band}
\alias{select_band}
\title{Bandwidth selection procedure for kernel spectral density estimator}
\usage{
select_band(
  ppp,
  inten.formula = NULL,
  data.covariate = NULL,
  a = 0.025,
  band.range,
  correct = TRUE,
  A1 = NULL,
  A2 = A1,
  endpt = 1.5,
  equal = TRUE,
  kern = bartlett_uni
)
}
\arguments{
\item{ppp}{A point pattern of class \code{"ppp"}.}

\item{inten.formula}{A \code{\link{formula}} syntax in character format specifying the
log-liner model for the intensity function, which is passed to \code{\link[=ppm]{ppm()}}. The
default is constant intensity \code{inten.formula = "~1"}.}

\item{data.covariate}{Optional. The values of spatial covariates passed to
the \code{data} argument in \code{\link[=ppm]{ppm()}}.}

\item{a}{Taper coefficient, a value within \eqn{(0,1/2)}. If \code{a = 0}, then
taper is not applied.}

\item{band.range}{Numeric vector. Search space for the optimal bandwidth.}

\item{correct}{Logical. If \code{TRUE} (default), conduct edge correction when
computing the kernel spectral estimator.}

\item{A1, A2}{Optional. Side lengths of the observation window.}

\item{endpt}{A positive value indicating the scale factor of the endpoint
frequency.}

\item{equal}{Logical. Whether to use the same bandwidth for both x and y axis.
The default is \code{TRUE}.}

\item{kern}{Univariate scaled kernel function. The default is Barlett kernel.}
}
\value{
A list.
}
\description{
Bandwidth selection procedure for kernel spectral density estimator
}
\examples{
library(spatstat)
lam <- function(x, y, m) {(x^2 + y) * ifelse(m == "A", 2, 1)}
set.seed(227823)
spp <- rmpoispp(lambda = lam, win = square(5), types = c("A","B"))
b <- seq(0.95, 1.4, 0.05) # You may use finer grid to search optimal bandwidth

cv <- select_band(spp, inten.formula = "~ x + y", band.range = b)
cv$OptimalBandwidth # Print the optimal bandwidth
plot(cv$Result[1,], cv$Result[2,], type = "b", pch = 20, las = 1,
     xlab = "Bandwidth", ylab = "Spectral divergence")
abline(v = cv$OptimalBandwidth, col = "blue")
}
