% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancillary.R
\name{taper}
\alias{taper}
\title{Univariate data taper}
\usage{
taper(x, a)
}
\arguments{
\item{x}{Cartesian coordinate of events. This argument is vectorized.}

\item{a}{A value in unit interval.}
}
\description{
Univariate data taper
}
