% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bartlett_uni.R
\name{bartlett_uni}
\alias{bartlett_uni}
\title{Scaled Battlett kernel (univariate)}
\usage{
bartlett_uni(v, b)
}
\arguments{
\item{v}{Input value.}

\item{b}{Bandwidth.}
}
\description{
Scaled Battlett kernel (univariate)
}
