% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coherence.R
\name{coherence}
\alias{coherence}
\title{Maximum coherence and partial coherence matrix}
\usage{
coherence(sp.est, ppp, type = "partial", all = FALSE)
}
\arguments{
\item{sp.est}{List. The kernel spectral estimate from \code{\link[=periodogram_smooth]{periodogram_smooth()}}.}

\item{ppp}{A point pattern of class \code{"ppp"}.}

\item{type}{If \code{"partial"} (default), calculate the partial coherence. If \code{"normal"}, calculate the coherence.}

\item{all}{Logical. If \code{TRUE}, extract the maximum (partial) coherence across all frequencies, which is not recommended.}
}
\value{
A matrix with each entry storing the maximum (partial) coherence of the two individual point processes.
}
\description{
Computes the coherence or partial coherence matrix for a multivatiate (multitype) point pattern.
}
\examples{
library(spatstat)
lam <- function(x, y, m) {(x^2 + y) * ifelse(m == "A", 2, 1)}
spp <- rmpoispp(lambda = lam, win = square(5), types = c("A","B"))

# Compute kernel spectral estimator with fitted intensity by log-linear model:
# with Cartesian coordinates
spectra <- periodogram_smooth(spp, inten.formula = "~ x + y", bandwidth = 1.2)
coh.partial <- coherence(spectra, spp) # Compute the maximum partial coherence
attr(coh.partial, "CohTable") # Print the partial coherence values for all frequencies

}
