% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hmatrix.R
\name{Hmatrix}
\alias{Hmatrix}
\title{Compute the the H matrix used in coherence.R}
\usage{
Hmatrix(sp.est, ppp)
}
\arguments{
\item{sp.est}{A list of kernel spectral estimate matrices of the point pattern.}

\item{ppp}{Point pattern.}
}
\description{
Compute the the H matrix used in coherence.R
}
