% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRspec.R
\name{IRspec}
\alias{IRspec}
\title{Spectrum of the intensity reweighted process}
\usage{
IRspec(w1, w2, sp.est, i = NULL, j = NULL, H.list = NULL, ppp)
}
\arguments{
\item{w1, w2}{Frequency vector (only allow frequency values evaluated in \code{sp.est})}

\item{sp.est}{A list of kernel spectral estimate matrices of the point pattern.}

\item{i, j}{Optional. Index of the multivariate point process.}

\item{H.list}{Optional. A list from \code{\link[=Hmatrix]{Hmatrix()}}.}

\item{ppp}{Point pattern.}
}
\value{
If both \code{i} and \code{j} are not specified (default), return a \eqn{m
\times m} matrix. If they are specified, return a value.
}
\description{
This function computes the estimator of the spectrum for a intensity reweighted
point process \eqn{\widehat{\mathcal{F}^{-1}(L_2)}(\boldsymbol\omega)}, which
is used to calculate the coherence and partial coherence.
\deqn{\widehat{\mathcal{F}^{-1}(L_2)}(\boldsymbol\omega) = H_{h,2}\left(\hat{
F}_b(\boldsymbol\omega)-(2\pi)^{-2}H^{-1}_{h,2}\text{diag}(H_{h^2\underline{\hat
\lambda}})\right)\oslash H_{h^2\underline{\hat\lambda}\cdot\underline{\hat
\lambda}^\intercal}}
}
