% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual.R
\name{plot_coher}
\alias{plot_coher}
\title{Plot coherence and partial coherence}
\usage{
plot_coher(sp.est, coh.mat, partial.coh.mat, xnorm = TRUE, ylim = NULL)
}
\arguments{
\item{sp.est}{List. The kernel spectral density estimate from \code{\link[=periodogram_smooth]{periodogram_smooth()}}.}

\item{coh.mat}{Coherence matrix from \code{\link[=coherence]{coherence()}} with \code{type = "normal"}.}

\item{partial.coh.mat}{Partial coherence matrix from \code{\link[=coherence]{coherence()}} with \code{type = "partial"}.}

\item{xnorm}{Logical. If \code{TRUE} (default), plot the radial averaged values.
If \code{FALSE}, plot the raw coherence and partial coherence values via heatmap.}

\item{ylim}{A numeric vector \code{c(lower, upper)} to specify the range to draw
for the radially-averaged plot. Not required if \code{xnorm = FALSE}.}
}
\description{
Use \code{ggplot2} package to visualize the coherence and partial coherence.
}
\examples{
library(spatstat)
lam <- function(x, y, m) {(x^2 + y) * ifelse(m == "A", 2, 1)}
set.seed(227823)
spp <- rmpoispp(lambda = lam, win = square(5), types = c("A","B"))

KSDE.list <- periodogram_smooth(spp, inten.formula = "~ x + y", bandwidth = 1.15)
coh.partial <- coherence(KSDE.list, spp)
coh <- coherence(KSDE.list, spp, type = "normal")
plot_coher(KSDE.list, coh, coh.partial)
}
\seealso{
\code{\link[=coherence]{coherence()}}
}
