% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoother.R
\name{smoother}
\alias{smoother}
\title{Smooth an element from a matrix by averaging its neighbors}
\usage{
smoother(w, period.mat, w.k1, w.k2, b1 = 1, b2 = b1, loo = FALSE, kernel_uni)
}
\arguments{
\item{w}{The single frequency to be smoothed.}

\item{period.mat}{The whole matrix, could be complex-valued.}

\item{w.k1, w.k2}{Vectors containing the whole frequencies at horizontal and
vertical directions.}

\item{b1, b2}{Bandwidths at horizontal and vertical direction, respectively.}

\item{loo}{Logical. If \code{TRUE}, conduct leave-one-out kernel smoothing (the
center \code{w} will be excluded when averaging). Otherwise, keep the center \code{w}
for averaging.}

\item{kernel_uni}{Univariate kernel function.}
}
\description{
Smooth an element from a matrix by averaging its neighbors
}
